Add Dockerfile for uploading artifact (#11)
Adds a `Dockerfile` which allows for building & uploading the
`protoc-gen-connect-swift` plugin to the BSR.

For now, this is being pushed to my personal BSR account (though it's
public). Once we are ready to OSS the repo, it'll be moved to the
official Buf plugins. The reason for this delay is that once we make it
an official plugin, it'll be publicly discoverable at
https://buf.build/plugins.

Co-authored-by: srikrsna <skrishna@buf.build>