Enforce license header validation on CI & update files (#38)
Adds integration with Buf's `licenseheaders` tool and enforces that each
file has the appropriate license headers on CI. Also updates the Swift
PM license template for the project so that it matches this new format.

Note that `Package.swift` is excluded because Swift requires that the
file starts with `// swift-tools-version`. Additionally, generated
outputs are omitted.