Update mocks to delay output streams until request is sent (#49)
In production, responses are only received over streams once a request
has been sent. This PR updates the included mocks to reflect this
behavior (rather than immediately returning all mocked responses as soon
as the stream is invoked).

This behavior is powered by `Combine`, making the `inputs` observable to
the mocks (and consumers).

The async/await implementations observe `$inputs` and then publish
responses once one input/request is sent.

The callback implementations also await `$inputs`, though this is done
in a one-liner in the generated mocks since they do not expose the
`results` themselves and instead invoke `onResult`.