Convert `ResponseMessage<T>` to `Result<ResponseMessage<T>, E>`
This has the benefit of ensuring that only one of `response` or `error` is specified, making it clear to consumers that they are mutually exclusive. This also adds helper extensions to the relevant `Result` to provide `.response` and `.error` accessors so that consumers don't necessarily have to `switch` over the result.