Add tracing info with HTTP status to `HTTPResponse` (#84)
It was pointed out that this information can be helpful for tracing,
logging, etc., so we are exposing it on the `HTTPResponse` to
interceptors and HTTP client subclasses. Because the HTTP status can
change based on the protocol in use, it is being exposed through a
separate `TracingInfo` type to make this more explicit.

Note that codes from `Error` types that are returned by `URLSession`
(for example, when the device is offline) are not passed through since
these are not valid HTTP statuses (usually negative numbers).