Update crosstest reference text (#163)
This updates references to the crosstest repo to use conformance
instead.

Note that the Docker image used for crosstests still uses an old
bufbuild image. This is because the test proto packages have changed in
the new repo and switching to use one of the new images would cause all
existing code to break. This will be fixed in a follow-up PR.