Use Conformance repo proto files (#165)
This adds some improvements and updates to the Connect Conformance
Tests. Namely:

* Use the [proto
definitions](https://github.com/connectrpc/conformance/tree/main/proto)
in the Conformance repo rather than house local files.
* Add an additional test that validates errors after response. The
conformance implementation for `failServerStreaming` changed
[here](https://github.com/connectrpc/conformance/pull/386) to inspect
response parameters and return that amount of responses before failing
so that users could test trailers-only responses. This fixes the current
test so that it doesn't send any response parameters and expects an
error straightaway. In addition, it adds another test to test that some
messages are received before the error.

Note that this updates Connect-Swift to use commit
[6ce5b0](https://github.com/connectrpc/conformance/commit/36ce5b0ce50bc8cb3314eee08a72eeda4c96721e),
but this is not the current commit. This is because all commits
afterwards cause two additional tests to fail (`failUnary` and
`failServerStreaming`) due to issues with parsing ErrorDetails. Those
will be fixed in a follow-up PR.

---------

Co-authored-by: Michael Rebello <me@michaelrebello.com>