Update to latest conformance image (#166)
This updates the repo to use the latest conformance image, which uses
the new organization on Dockerhub as well as the new proto packages
(`connectrpc.conformance.v1` instead of `grpc.testing`).

As a part of this update, this attempts a fix at decoding the
`ErrorDetails` value, which is returned as a base64-encoded string. The
Connect protocol states that implementations should emit unpadded values
and that downstream implementations should accept both padded and
unpadded values. This adds padding to this string if needed as part of
the `ConnectError` decode logic.

---------

Co-authored-by: Michael Rebello <me@michaelrebello.com>