Add test for deserializing unpadded base64 (#177)
This adds a test for deserializing an `ErrorDetail` value with unpadded
base64. By default, Swift pads base64-encoding and it seems the only way
to not pad is to manually remove the `=`. This adds an option to the
`errorData` utility function to remove padding so both scenarios can be
tested.