Revert usage of `package` keyword & unsafe flags (#218)
Resolves https://github.com/connectrpc/connect-swift/issues/217.

Instead of using the `package` keyword which is supported by Swift 5.9+
but does not seem to work out-of-the-box because I didn't see
`-package-name` get passed by Xcode to the build system, I'm reverting
back to just keeping these few files in a `PackageInternal` directory
and documenting them as not being part of the public interface. This
eliminates the need for unsafe flags and resolves the aforementioned
issue.