Update conformance runner & tests to v1.0.1 (#247)
Updates the Connect conformance runner and test suite to
[v1.0.1](https://github.com/connectrpc/conformance/releases/tag/v1.0.1)
and fixes a few failures from new tests that were added. It also opts
out of a few edge cases which are not yet fully specced out / required
by the Connect protocol but are now enforced by the conformance runner
for the sake of consistency. We may implement logic to conform to these
in the future.