Add package visibility option to code generators (#355)
The `package` access modifier was introduced in Swift 5.9 ([proposal
0386](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0386-package-access-modifier.md))
to bridge the gap between `internal` and `public` in Swift packages.

[Support for this modifier in generated
code](https://github.com/apple/swift-protobuf/pull/1472) was added to
SwiftProtobuf in September 2023.

This change adds the `Package` visibility option to the Connect
generators. This is particularly useful when developing a Swift package,
and you want to vend your generated models and mocks from a specific
target to other targets of your package but not to dependent
applications.

---------

Signed-off-by: Eddie Seay <eddie.seay@cfacorp.com>